% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallbackAsyncFSelect.R
\name{assert_async_fselect_callback}
\alias{assert_async_fselect_callback}
\alias{assert_async_fselect_callbacks}
\title{Assertions for Callbacks}
\usage{
assert_async_fselect_callback(callback, null_ok = FALSE)

assert_async_fselect_callbacks(callbacks)
}
\arguments{
\item{callback}{(\link{CallbackAsyncFSelect}).}

\item{null_ok}{(\code{logical(1)})\cr
If \code{TRUE}, \code{NULL} is allowed.}

\item{callbacks}{(list of \link{CallbackAsyncFSelect}).}
}
\value{
[CallbackAsyncFSelect | List of \link{CallbackAsyncFSelect}s.
}
\description{
Assertions for \link{CallbackAsyncFSelect} class.
}
