% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_fselect.R
\name{ensemble_fselect}
\alias{ensemble_fselect}
\title{Ensemble Feature Selection}
\usage{
ensemble_fselect(
  fselector,
  task,
  learners,
  init_resampling,
  inner_resampling,
  measure,
  terminator,
  callbacks = list(),
  store_models = TRUE
)
}
\arguments{
\item{fselector}{(\link{FSelector})\cr
Optimization algorithm.}

\item{task}{(\link[mlr3:Task]{mlr3::Task})\cr
Task to operate on.}

\item{learners}{(list of \link[mlr3:Learner]{mlr3::Learner})\cr
The learners to be used for feature selection.}

\item{init_resampling}{(\link[mlr3:Resampling]{mlr3::Resampling})\cr
The initial resampling strategy of the data, from which each train set
will be passed on to the learners.
Can only be \link{mlr_resamplings_subsampling} or \link{mlr_resamplings_bootstrap}.}

\item{inner_resampling}{(\link[mlr3:Resampling]{mlr3::Resampling})\cr
The inner resampling strategy used by the \link{FSelector}.}

\item{measure}{(\link[mlr3:Measure]{mlr3::Measure})\cr
Measure to optimize. If \code{NULL}, default measure is used.}

\item{terminator}{(\link{Terminator})\cr
Stop criterion of the feature selection.}

\item{callbacks}{(list of \link{CallbackFSelect})\cr
List of callbacks.}
}
\description{
Ensemble feature selection using multiple learners.
The ensemble feature selection method is designed to identify the
most informative features from a given dataset by leveraging multiple
machine learning models and resampling techniques.
}
\details{
The method begins by applying an initial resampling technique specified
by the user, to create multiple subsamples from the original dataset.
This resampling process helps in generating diverse subsets of data for
robust feature selection.
For each subsample generated in the previous step, the method performs
wrapped-based feature selection using each provided learner, an inner
resampling method and a performance measure.
This process generates a best feature subset for each combination of
subsample and learner.
All the results are stored in a \link{data.table} object.
}
\examples{
\donttest{

  ensemble_fselect(
    fselector = fs("random_search"),
    task = tsk("sonar"),
    learners = lrns(c("classif.rpart", "classif.featureless")),
    init_resampling = rsmp("subsampling", repeats = 2),
    inner_resampling = rsmp("cv", folds = 3),
    measure = msr("classif.ce"),
    terminator = trm("evals", n_evals = 10)
  )
}
}
