% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\docType{data}
\name{FeatureSelection}
\alias{FeatureSelection}
\title{Feature Selection Wrapper Base Class}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
This is the base class for feature selection wrappers.
}
\section{Construction}{
\preformatted{fs = FeatureSelectionr$new(id, pe, tm, measure, param_set, param_vals)
}
\itemize{
\item \code{id} :: \code{character(1)}\cr Identifier for the wrapper.
\item \code{pe} :: \link{PerformanceEvaluator}\cr
\item \code{tm} :: \link{Terminator}\cr
\item \code{param_set} :: \link[paradox:ParamSet]{paradox::ParamSet}\cr Set of hyperparameters.
\item \code{param_vals} :: named \code{list()}\cr Named list of hyperparameter settings.
}
}

\section{Fields}{

\itemize{
\item \code{id} :: \code{character(1)}\cr Stores the identifier of of the wrapper.
\item \code{pe} :: \link{PerformanceEvaluator}\cr The current state of the \link{PerformanceEvaluator}.
\item \code{tm} :: \link{Terminator}\cr The current state of the \link{Terminator}.
\item \code{measure} :: \link[mlr3:Measure]{mlr3::Measure}\cr Stores the measure.
\item \code{param_set} :: \link[paradox:ParamSet]{paradox::ParamSet}\cr Description of available hyperparameters and hyperparameter settings.
\item \code{state} :: list of \code{integer()}\cr Stores the currently evaluated feature sets in 0/1 encodings.
}
}

\section{Methods}{

\itemize{
\item \code{$calculate()}\cr Performs the feature selection until the budget of \link{Terminator} is exhausted.
}
}

\section{Parameter Values}{

\itemize{
\item \code{max_features} :: \code{integer(1)}\cr Maximum number of features in set.
}
}

\seealso{
Other FeatureSelection: \code{\link{FeatureSelectionExhaustive}},
  \code{\link{FeatureSelectionGenetic}},
  \code{\link{FeatureSelectionRandom}},
  \code{\link{FeatureSelectionSequential}}
}
\concept{FeatureSelection}
\keyword{datasets}
