% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelectionExhaustive.R
\docType{data}
\name{FeatureSelectionExhaustive}
\alias{FeatureSelectionExhaustive}
\title{FeatureSelectionExhaustive}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{FeatureSelection}.}
\usage{
FeatureSelectionExhaustive
}
\description{
Exhaustive feature selection wrapper.
}
\section{Construction}{
\preformatted{fs = FeatureSelectionExhaustive$new(pe, tm, measure, param_vals)
}

For arguments, see \link{FeatureSelection}.
}

\section{Fields}{

See \link{FeatureSelection}.
}

\section{Methods}{

See \link{FeatureSelection} and additionally:
\itemize{
\item \code{$get_result()}\cr Returns best feature set.
}
}

\section{Parameter Values}{

\itemize{
\item \code{max_features} (\code{integer(1)})\cr Maximum number of features in set.
}
}

\examples{
task = mlr3::mlr_tasks$get("pima")
measure = mlr3::mlr_measures$get(c("classif.acc"))
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("cv", param_vals = list(folds = 5L))
resampling$instantiate(task)
pe = PerformanceEvaluator$new(task = task, learner = learner, resampling = resampling)
tm = TerminatorRuntime$new(max_time = 20, units = "secs")
fs = FeatureSelectionExhaustive$new(pe = pe, tm = tm, measure,
                                    param_vals = list(max_features = 3))
fs$calculate()
fs$get_result()
}
\seealso{
Other FeatureSelection: \code{\link{FeatureSelectionGenetic}},
  \code{\link{FeatureSelectionRandom}},
  \code{\link{FeatureSelectionSequential}},
  \code{\link{FeatureSelection}}
}
\concept{FeatureSelection}
\keyword{datasets}
