% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelectionForward.R
\name{FeatureSelectionForward}
\alias{FeatureSelectionForward}
\title{FeatureSelectionForward}
\description{
FeatureSelection child class to conduct forward search.
}
\section{Usage}{
\preformatted{fs = FeatureSelectionForward$new()
}

See \link{FeatureSelection} for a description of the interface.
}

\section{Arguments}{

\itemize{
\item \code{pe} (\code{[PerformanceEvaluator]}).
\item \code{tm} (\code{[Terminator]}).
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{FeatureSelectionForward}.
\code{$get_result()} Returns selected features in each step.
The interface is described in \link{FeatureSelection}.

Each step is possibly executed in parallel via \code{\link[mlr3:benchmark]{mlr3::benchmark()}}
}

\examples{
task = mlr3::mlr_tasks$get("pima")
measures = mlr3::mlr_measures$mget(c("classif.acc"))
task$measures = measures
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("cv", param_vals = list(folds = 5L))
pe = PerformanceEvaluator$new(task, learner, resampling)
tm = TerminatorPerformanceStep$new(threshold = 0.01)
fs = FeatureSelectionForward$new(pe, tm)
fs$calculate()
fs$get_result()
}
\seealso{
Other FeatureSelection: \code{\link{FeatureSelectionGenetic}},
  \code{\link{FeatureSelectionRandom}},
  \code{\link{FeatureSelection}}
}
\concept{FeatureSelection}
