% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelectionGenetic.R
\docType{data}
\name{FeatureSelectionGenetic}
\alias{FeatureSelectionGenetic}
\title{FeatureSelectionGenetic}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{FeatureSelection}.}
\usage{
FeatureSelectionGenetic
}
\description{
Genetic feature selection wrapper. The comma strategy \code{(mu, lambda)} selects a new population of size \code{mu} out of the \code{lambda > mu} offspring. The plus strategy \code{(mu + lambda)} uses the joint pool of \code{mu} parents and \code{lambda} offspring for selecting \code{mu} new candidates. Out of those \code{mu} features, the new \code{lambda} features are generated by randomly choosing pairs of parents. These are crossed over and \code{crossover_rate} represents the probability of choosing a feature from the first parent instead of the second parent. The resulting offspring is mutated, i.e., its bits are flipped with probability \code{mutation_rate}.
}
\section{Construction}{
\preformatted{fs = FeatureSelectionGenetic$new(pe, tm, measure, param_vals)
}

For arguments, see \link{FeatureSelection}.
}

\section{Fields}{

See \link{FeatureSelection}.
}

\section{Methods}{

See \link{FeatureSelection} and additionally:
\itemize{
\item \code{$get_result()}\cr Returns best feature set.
\item \code{$get_path()}\cr Returns each generation.
}
}

\section{Parameter Values}{

See \link{FeatureSelection} and additionally:
\itemize{
\item \code{mu} :: \code{integer(1)}\cr Size of the parent population.
\item \code{lambda} :: \code{integer(1)}\cr Size of the children population.
\item \code{crossover_rate} :: \code{numeric(1)}\cr Probability of choosing a feature from the first parent within the crossover mutation.
\item \code{mutation_rate} :: \code{numeric(1)}\cr Probability of flipping a feature, i.e. switch between selecting / deselecting a feature.
\item \code{max_features} :: \code{integer(1)}\cr Maximum number of features.
\item \code{strategy} :: \code{character(1)}\cr \code{plus} or \code{comma}. Indicates whether to use a \code{(mu, lambda)} or \code{(mu + lambda)} genetic algorithm.
}
}

\examples{
task = mlr3::mlr_tasks$get("pima")
measures = mlr3::mlr_measures$get(c("classif.acc"))
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("cv", param_vals = list(folds = 5L))
resampling$instantiate(task)
pe = PerformanceEvaluator$new(task, learner, resampling)
tm = TerminatorRuntime$new(max_time = 10, units = "secs")
fs = FeatureSelectionGenetic$new(pe, tm, measures,
                                 param_vals = list(mu = 10,
                                                   lambda = 20,
                                                   strategy = "comma"))
fs$calculate()
fs$get_result()
fs$get_path()
}
\seealso{
Other FeatureSelection: \code{\link{FeatureSelectionExhaustive}},
  \code{\link{FeatureSelectionRandom}},
  \code{\link{FeatureSelectionSequential}},
  \code{\link{FeatureSelection}}
}
\concept{FeatureSelection}
\keyword{datasets}
