% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelectionSequential.R
\docType{data}
\name{FeatureSelectionSequential}
\alias{FeatureSelectionSequential}
\title{FeatureSelectionSequential}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{FeatureSelection}.}
\usage{
FeatureSelectionSequential
}
\description{
Sequential feature selection wrapper.
}
\section{Construction}{
\preformatted{fs = FeatureSelectionSequential$new(pe, tm, measure, param_vals)
}

For arguments, see \link{FeatureSelection}.
}

\section{Fields}{

See \link{FeatureSelection}.
}

\section{Methods}{

See \link{FeatureSelection} and additionally:
\itemize{
\item \code{$get_result()}\cr Returns best feature set.
\item \code{$get_path()}\cr Returns each step.
}
}

\section{Parameter Values}{

\itemize{
\item \code{max_features} (\code{integer(1)})\cr Maximum number of features in set.
\item \code{strategy} (\code{character(1)})\cr \code{fsf} for forward or \code{fsb} for backward feature selection.
}
}

\examples{
task = mlr3::mlr_tasks$get("pima")
measure = mlr3::mlr_measures$get(c("classif.acc"))
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("cv", param_vals = list(folds = 5L))
resampling$instantiate(task)
pe = PerformanceEvaluator$new(task, learner, resampling)
tm = TerminatorPerformanceStep$new(threshold = 0.01)
fs = FeatureSelectionSequential$new(pe = pe, tm = tm, measure,
                                    param_vals = list(max_features = 4))
fs$calculate()
fs$get_result()
}
\seealso{
Other FeatureSelection: \code{\link{FeatureSelectionExhaustive}},
  \code{\link{FeatureSelectionGenetic}},
  \code{\link{FeatureSelectionRandom}},
  \code{\link{FeatureSelection}}
}
\concept{FeatureSelection}
\keyword{datasets}
