% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{Filter}
\alias{Filter}
\title{Filter Base Class}
\description{
FilterBase.
}
\section{Usage}{
\preformatted{Filter = Filter$new(id)
# public members
filter$id
filter$ff
filter$settings
filter$values
# public methods
filter$tune()
filter$tune_result()
}
}

\section{Arguments}{

\itemize{
\item \code{id} (\code{character(1)}):\cr
The id of the Filter
\item \code{packages} (\code{character()}]:\cr
Set of required packages.
\item \code{supported_features} (\code{list}):\cr
The feature types supported by the filter.
\item \code{supported_tasks} (\code{list}):\cr
The task types supported by the filter.
\item \code{settings} (\code{list}):\cr
The settings for the Filter
}
}

\section{Details}{

\itemize{
\item \code{$calculate(task)} calculates the filter values.
\item \code{$new()} creates a new object of class \link{Filter}.
\item \code{$filter()} filters a \link{Task} using specific criteria
\itemize{
\item returns a subsetted \link{Task}
}
\item \code{$id} stores an identifier for this \link{Filter}.
\item \code{$filter_values} stores the calculated filter values.
\item \code{$packages} stores the names of required packages.
\item \code{$settings} is a list of hyperparamter settings for this \link{Filter}.
}
}

\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterDISR}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}},
  \code{\link{FilterLinearCorrelation}},
  \code{\link{FilterMIM}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterRankCorrelation}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{mlr_filters}}
}
\concept{Filter}
