% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterAUC.R
\docType{data}
\name{FilterAUC}
\alias{FilterAUC}
\alias{mlr_filters_auc}
\title{AUC Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{FilterResult}.}
\usage{
FilterAUC
}
\description{
Area under the (ROC) Curve filter.
Calls \code{\link[Metrics:auc]{Metrics::auc()}}, and scores with the absolute value of the difference between the AUC and 0.5.
}
\examples{
task = mlr3::mlr_tasks$get("sonar")
filter = FilterAUC$new()
filter$calculate(task)
as.data.table(filter)[1:3]
}
\seealso{
Other Filter: \code{\link{FilterCMIM}},
  \code{\link{FilterDISR}}, \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}},
  \code{\link{FilterLinearCorrelation}},
  \code{\link{FilterMIM}}, \code{\link{FilterMRMR}},
  \code{\link{FilterNJMIM}},
  \code{\link{FilterRankCorrelation}},
  \code{\link{FilterResult}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
