% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCarScore.R
\docType{data}
\name{FilterCarScore}
\alias{FilterCarScore}
\alias{mlr_filters_carscore}
\title{Conditional Mutual Information Based Feature Selection Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterCarScore
}
\description{
Filter \code{carscore} determines the Correlation-Adjusted (marginal) coRelation
scores (short CAR scores). The CAR scores for a set of features are defined as the
correlations between the target and the decorrelated features.
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterDISR}},
  \code{\link{FilterEmbedded}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}},
  \code{\link{FilterLinearCorrelation}},
  \code{\link{FilterMIM}}, \code{\link{FilterMRMR}},
  \code{\link{FilterNJMIM}},
  \code{\link{FilterRankCorrelation}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
