% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCMIM.R
\docType{data}
\name{FilterCMIM}
\alias{FilterCMIM}
\alias{mlr_filters_cmim}
\title{Minimal Conditional Mutual Information Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterCMIM
}
\description{
Minimal conditional mutual information maximisation filter.
Calls \code{\link[praznik:CMIM]{praznik::CMIM()}}.
}
\examples{
task = mlr3::mlr_tasks$get("iris")
filter = FilterCMIM$new()
filter$calculate(task)
as.data.table(filter)[1:3]
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCarScore}}, \code{\link{FilterDISR}},
  \code{\link{FilterEmbedded}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}},
  \code{\link{FilterLinearCorrelation}},
  \code{\link{FilterMIM}}, \code{\link{FilterMRMR}},
  \code{\link{FilterNJMIM}},
  \code{\link{FilterRankCorrelation}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
