% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCorrelation.R
\docType{data}
\name{FilterCorrelation}
\alias{FilterCorrelation}
\alias{mlr_filters_correlation}
\title{Correlation Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterCorrelation
}
\description{
Correlation filter. Calls \code{\link[stats:cor]{stats::cor()}}.
}
\examples{
task = mlr3::mlr_tasks$get("mtcars")
filter = FilterCorrelation$new()
filter$calculate(task)
as.data.table(filter)[1:3]
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterCarScore}},
  \code{\link{FilterDISR}}, \code{\link{FilterEmbedded}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}}, \code{\link{FilterMIM}},
  \code{\link{FilterMRMR}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
