% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterGainRatio.R
\docType{data}
\name{FilterGainRatio}
\alias{FilterGainRatio}
\alias{mlr_filters_gain_ratio}
\title{Gain Ratio Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterGainRatio
}
\description{
Gain Ratio filter.
Calls \code{\link[FSelectorRcpp:information_gain]{FSelectorRcpp::information_gain()}}.
The target variable of regression tasks is automatically binned (argument \code{equal} in \code{\link[FSelectorRcpp:information_gain]{FSelectorRcpp::information_gain()}}).
}
\examples{
task = mlr3::mlr_tasks$get("sonar")
filter = FilterGainRatio$new()
filter$calculate(task)
head(as.data.table(filter), 3)
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterDISR}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}},
  \code{\link{FilterLinearCorrelation}},
  \code{\link{FilterMIM}}, \code{\link{FilterMRMR}},
  \code{\link{FilterNJMIM}},
  \code{\link{FilterRankCorrelation}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariableImportance}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
