% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterKruskalTest.R
\docType{data}
\name{FilterKruskalTest}
\alias{FilterKruskalTest}
\alias{mlr_filters_kruskal_test}
\title{Kruskal-Wallis Test Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterKruskalTest
}
\description{
Kruskal-Wallis rank sum test filter. Calls
\code{\link[stats:kruskal.test]{stats::kruskal.test()}}.
}
\examples{
task = mlr3::mlr_tasks$get("iris")
filter = FilterKruskalTest$new()
filter$calculate(task)
as.data.table(filter)[1:3]
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterCarScore}},
  \code{\link{FilterCorrelation}},
  \code{\link{FilterDISR}}, \code{\link{FilterEmbedded}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterMIM}}, \code{\link{FilterMRMR}},
  \code{\link{FilterNJMIM}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
