% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterMRMR.R
\docType{data}
\name{FilterMRMR}
\alias{FilterMRMR}
\alias{mlr_filters_MRMR}
\title{Minimum redundancy maximal relevancy filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{FilterResult}.}
\usage{
FilterMRMR
}
\description{
Minimum redundancy maximal relevancy filter.
Calls \code{\link[praznik:MRMR]{praznik::MRMR()}}.
}
\examples{
task = mlr3::mlr_tasks$get("iris")
filter = FilterMRMR$new()
filter$calculate(task)
head(as.data.table(filter), 3)
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterCarScore}},
  \code{\link{FilterDISR}}, \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}},
  \code{\link{FilterLinearCorrelation}},
  \code{\link{FilterMIM}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterRankCorrelation}},
  \code{\link{FilterResult}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
