% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterRankCorrelation.R
\docType{data}
\name{FilterRankCorrelation}
\alias{FilterRankCorrelation}
\alias{mlr_filters_rank_correlation}
\title{Rank Correlation Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterRankCorrelation
}
\description{
Rank correlation filter.
Calls \code{\link[stats:cor]{stats::cor()}}.
}
\examples{
task = mlr3::mlr_tasks$get("mtcars")
filter = FilterRankCorrelation$new()
filter$calculate(task)
head(as.data.table(filter), 3)
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterDISR}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}},
  \code{\link{FilterLinearCorrelation}},
  \code{\link{FilterMIM}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariableImportance}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
