% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterVariableImportance.R
\docType{data}
\name{FilterVariableImportance}
\alias{FilterVariableImportance}
\alias{mlr_filters_variable_importance}
\title{Variable Importance Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{FilterResult}.}
\usage{
FilterVariableImportance
}
\description{
Variable Importance filter.
Takes a \link[mlr3:Learner]{mlr3::Learner} which supports retrieving the variable importance (property "importance"),
fits the model and uses the importance values as filter scores.
}
\examples{
task = mlr3::mlr_tasks$get("iris")
learner = mlr3::mlr_learners$get("classif.rpart")
filter = FilterVariableImportance$new(learner = learner)
filter$calculate(task)
as.data.table(filter)[1:3]
}
\concept{FilterResult}
\keyword{datasets}
