% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterVariance.R
\docType{data}
\name{FilterVariance}
\alias{FilterVariance}
\alias{mlr_filters_variance}
\title{Variance Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterVariance
}
\description{
Variance filter.
Calls \code{\link[stats:var]{stats::var()}}. Argument \code{na.rm} defaults to \code{TRUE} here.
}
\examples{
task = mlr3::mlr_tasks$get("mtcars")
filter = FilterVariance$new()
filter$calculate(task)
as.data.table(filter)[1:3]
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterCarScore}},
  \code{\link{FilterCorrelation}},
  \code{\link{FilterDISR}}, \code{\link{FilterEmbedded}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}}, \code{\link{FilterMIM}},
  \code{\link{FilterMRMR}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{Filter}}, \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
