% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSelectEvolutionary.R
\name{FSelectEvolutionary}
\alias{FSelectEvolutionary}
\title{FSelectEvolutionary Class}
\description{
Class for evolutionary feature selection calling \code{\link[ecr:ecr]{ecr::ecr()}} from package \CRANpkg{ecr}.
}
\section{Parameters}{

\describe{
\item{\code{mu}}{\code{integer(1)}}
\item{\code{lambda}}{\code{integer(1)}}
\item{\code{p.recomb}}{\code{double(1)}}
\item{\code{p.mut}}{\code{double(1)}}
\item{\code{survival.strategy}}{\code{character(1)}}
\item{\code{n.elite}}{\code{integer(1)}}
\item{\code{initial.solutions}}{list of \code{integer()}}
\item{\code{parent.selector}}{object of class \code{ecr::selector}}
\item{\code{survival.selector}}{object of class \code{ecr::selector}}}

For the meaning of the control parameter, see \code{\link[ecr:ecr]{ecr::ecr()}}.
Note that \code{mu} and \code{lambda} must be set by the user.
The \code{terminators} parameter is replaced by the \code{Terminator} subclasses.
}

\section{Super class}{
\code{\link[mlr3fselect:FSelect]{mlr3fselect::FSelect}} -> \code{FSelectEvolutionary}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FSelectEvolutionary$new()}}
\item \href{#method-clone}{\code{FSelectEvolutionary$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelect" data-id="format">}\href{../../mlr3fselect/html/FSelect.html#method-format}{\code{mlr3fselect::FSelect$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelect" data-id="print">}\href{../../mlr3fselect/html/FSelect.html#method-print}{\code{mlr3fselect::FSelect$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelect" data-id="select">}\href{../../mlr3fselect/html/FSelect.html#method-select}{\code{mlr3fselect::FSelect$select()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create new \code{FSelectEvolutionary } object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectEvolutionary$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{FSelectEvolutionary } object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectEvolutionary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
