% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSelectExhaustive.R
\name{FSelectExhaustive}
\alias{FSelectExhaustive}
\title{FSelectExhaustive Class}
\description{
Class for exhaustive feature selection. All feature sets are searched.
}
\section{Parameters}{

\describe{
\item{\code{max_features}}{\code{integer(1)} Maximum number of features. By default, number of features in \link[mlr3:Task]{mlr3::Task}.}}

In order to support general termination criteria and parallelization, feature sets are evaluated in batches.
The size of the feature sets is increased by 1 in each batch.
}

\examples{
library(mlr3)

terminator = term("evals", n_evals = 15)
instance = FSelectInstance$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measures = msr("classif.ce"),
  terminator = terminator
)

fs = fs("exhaustive")
fs$select(instance) # modifies the instance by reference
instance$result # returns best configuration and best performance
instance$archive() # allows access of data.table / benchmark result of full path of all evaluation
instance$optimization_path(n = 1, m = 1:4) # returns best feature set of each batch
}
\section{Super class}{
\code{\link[mlr3fselect:FSelect]{mlr3fselect::FSelect}} -> \code{FSelectExhaustive}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FSelectExhaustive$new()}}
\item \href{#method-clone}{\code{FSelectExhaustive$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelect" data-id="format">}\href{../../mlr3fselect/html/FSelect.html#method-format}{\code{mlr3fselect::FSelect$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelect" data-id="print">}\href{../../mlr3fselect/html/FSelect.html#method-print}{\code{mlr3fselect::FSelect$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelect" data-id="select">}\href{../../mlr3fselect/html/FSelect.html#method-select}{\code{mlr3fselect::FSelect$select()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create new \code{FSelectExhaustive} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectExhaustive$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{FSelectExhaustive}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectExhaustive$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
