% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSelectorEvolutionary.R
\name{FSelectorEvolutionary}
\alias{FSelectorEvolutionary}
\title{Feature Selection via Evolutionary Search}
\source{
\cite{mlr3fselect}{bossek_2017}
}
\description{
\code{FSelectEvolutionary} class that implements evolutionary search. Calls
\code{\link[ecr:ecr]{ecr::ecr()}} from package \CRANpkg{ecr}.
}
\note{
We add a random feature to the set if \code{ecr::ecr()} suggests an empty
feature set.
}
\section{Dictionary}{

This \link{FSelector} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary}
\link{mlr_fselectors} or with the associated sugar function \code{\link[=fs]{fs()}}:\preformatted{mlr_fselectors$get("evolutionary")
fs("evolutionary")
}
}

\section{Parameters}{

\describe{
\item{\code{mu}}{\code{integer(1)}}
\item{\code{lambda}}{\code{integer(1)}}
\item{\code{p.recomb}}{\code{double(1)}}
\item{\code{p.mut}}{\code{double(1)}}
\item{\code{survival.strategy}}{\code{character(1)}}
\item{\code{n.elite}}{\code{integer(1)}}
\item{\code{initial.solutions}}{list of \code{integer()}}
\item{\code{parent.selector}}{\code{ecr} selector}
\item{\code{survival.selector}}{\code{ecr} selector}
}

For the meaning of the control parameter, see \code{\link[ecr:ecr]{ecr::ecr()}}. Note that \code{mu}
and \code{lambda} must be set by the user. The \code{terminators} parameter is replaced
by the \link{Terminator} subclasses.
}

\examples{
library(mlr3)

terminator = trm("evals", n_evals = 10)

instance = FSelectInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = terminator
)

fselector = fs("evolutionary", mu = 10, lambda = 5)

# Modifies the instance by reference
fselector$optimize(instance)

# Returns best scoring evaluation
instance$result

# Allows access of data.table of full path of all evaluations
instance$archive$data
}
\section{Super class}{
\code{\link[mlr3fselect:FSelector]{mlr3fselect::FSelector}} -> \code{FSelectorEvolutionary}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FSelectorEvolutionary$new()}}
\item \href{#method-clone}{\code{FSelectorEvolutionary$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelector" data-id="format">}\href{../../mlr3fselect/html/FSelector.html#method-format}{\code{mlr3fselect::FSelector$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelector" data-id="optimize">}\href{../../mlr3fselect/html/FSelector.html#method-optimize}{\code{mlr3fselect::FSelector$optimize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelector" data-id="print">}\href{../../mlr3fselect/html/FSelector.html#method-print}{\code{mlr3fselect::FSelector$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectorEvolutionary$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectorEvolutionary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
