% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSelectorRFE.R
\name{FSelectorRFE}
\alias{FSelectorRFE}
\title{Feature Selection via Recursive Feature Elimination}
\description{
\code{FSelectorRFE} class that implements Recursive Feature Elimination (RFE). The
recursive algorithm (\code{recursive = TRUE}) recomputes the feature importance
on the reduced feature set in every iteration.  The non-recursive algorithm
(\code{recursive = FALSE}) only uses the feature importance of the model fitted
with all features to eliminate the next most unimportant features in every
iteration.
}
\section{Dictionary}{

This \link{FSelector} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary}
\link{mlr_fselectors} or with the associated sugar function \code{\link[=fs]{fs()}}:\preformatted{mlr_fselectors$get("rfe")
fs("rfe")
}
}

\section{Parameters}{

\describe{
\item{\code{min_features}}{\code{integer(1)}\cr
The minimum number of features to select, default is \code{1L}.}
\item{\code{feature_fraction}}{\code{double(1)}\cr
Fraction of features to retain in each iteration, default is \code{0.5}.}
\item{\code{feature_number}}{\code{integer(1)}\cr
Number of features to remove in each iteration.}
\item{\code{subset_sizes}}{\code{integer()}\cr
Vector of number of features to retain in each iteration. Must be sorted in
decreasing order.}
\item{\code{recursive}}{\code{logical(1)}\cr
Use the recursive version? Default is \code{FALSE}.}
}

The parameter \code{feature_fraction}, \code{feature_number} and \code{subset_sizes} are
mutually exclusive.
}

\examples{
library(mlr3)

terminator = trm("evals", n_evals = 10)
instance = FSelectInstanceSingleCrit$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = terminator,
  store_models = TRUE
)

fselector = fs("rfe")

# Modifies the instance by reference
fselector$optimize(instance)

# Returns best scoring evaluation
instance$result

# Allows access of data.table of full path of all evaluations
instance$archive$data()
}
\section{Super class}{
\code{\link[mlr3fselect:FSelector]{mlr3fselect::FSelector}} -> \code{FSelectorRFE}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{importance}}{\code{numeric()}\cr
Stores the feature importance of the model with all variables if
\code{recrusive} is set to \code{FALSE}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FSelectorRFE$new()}}
\item \href{#method-clone}{\code{FSelectorRFE$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelector" data-id="format">}\href{../../mlr3fselect/html/FSelector.html#method-format}{\code{mlr3fselect::FSelector$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelector" data-id="optimize">}\href{../../mlr3fselect/html/FSelector.html#method-optimize}{\code{mlr3fselect::FSelector$optimize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelector" data-id="print">}\href{../../mlr3fselect/html/FSelector.html#method-print}{\code{mlr3fselect::FSelector$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectorRFE$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectorRFE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
