% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3featsel-package}
\alias{mlr3featsel}
\alias{mlr3featsel-package}
\title{mlr3featsel: Feature Selection for 'mlr3'}
\description{
Implements methods for feature selection and
    filtering in 'mlr3'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/mlr3featsel}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3featsel/issues}
}

}
\author{
\strong{Maintainer}: Patrick Schratz \email{patrick.schratz@gmail.com} (0000-0003-0748-6624)

Authors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (0000-0001-6002-6980)
}

}
