% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3fselect-package}
\alias{mlr3fselect}
\alias{mlr3fselect-package}
\title{mlr3fselect: Feature selection for mlr3}
\description{
Implements methods for feature selection in mlr3.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3fselect.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3fselect}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3fselect/issues}
}

}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})

Authors:
\itemize{
  \item Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (\href{https://orcid.org/0000-0001-6002-6980}{ORCID})
}

}
