% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_callbacks.R
\name{mlr3fselect.async_freeze_archive}
\alias{mlr3fselect.async_freeze_archive}
\title{Freeze Archive Callback}
\description{
This \link{CallbackAsyncFSelect} freezes the \link{ArchiveAsyncFSelect} to \link{ArchiveAsyncFSelectFrozen} after the optimization has finished.
}
\examples{
clbk("mlr3fselect.async_freeze_archive")
}
