% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_callbacks.R
\name{mlr3fselect.internal_tuning}
\alias{mlr3fselect.internal_tuning}
\title{Internal Tuning Callback}
\description{
This callback runs internal tuning alongside the feature selection.
The internal tuning values are aggregated and stored in the results.
The final model is trained with the best feature set and the tuned value.
}
\examples{
clbk("mlr3fselect.internal_tuning")
}
