% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_callbacks.R
\name{mlr3fselect.one_se_rule}
\alias{mlr3fselect.one_se_rule}
\title{One Standard Error Rule Callback}
\source{
Kuhn, Max, Johnson, Kjell (2013).
\dQuote{Applied Predictive Modeling.}
In chapter Over-Fitting and Model Tuning, 61--92.
Springer New York, New York, NY.
ISBN 978-1-4614-6849-3.
}
\description{
Selects the smallest feature set within one standard error of the best as the result.
If there are multiple such feature sets with the same number of features, the first one is selected.
If the sets have exactly the same performance but different number of features,
the one with the smallest number of features is selected.
}
\examples{
clbk("mlr3fselect.one_se_rule")

# Run feature selection on the diabetes data set with the callback
instance = fselect(
  fselector = fs("random_search"),
  task = tsk("diabetes"),
  learner = lrn("classif.rpart"),
  resampling = rsmp ("cv", folds = 3),
  measures = msr("classif.ce"),
  term_evals = 10,
  callbacks = clbk("mlr3fselect.one_se_rule"))
# Smallest feature set within one standard error of the best
instance$result
}
