% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{mlr3fselect_assertions}
\alias{mlr3fselect_assertions}
\alias{assert_fselectors}
\alias{assert_fselector_async}
\alias{assert_fselector_batch}
\alias{assert_fselect_instance}
\alias{assert_fselect_instance_async}
\alias{assert_fselect_instance_batch}
\title{Assertion for mlr3fselect objects}
\usage{
assert_fselectors(fselectors)

assert_fselector_async(fselector)

assert_fselector_batch(fselector)

assert_fselect_instance(inst)

assert_fselect_instance_async(inst)

assert_fselect_instance_batch(inst)
}
\arguments{
\item{fselectors}{(list of \link{FSelector}).}

\item{fselector}{(\link{FSelectorBatch}).}

\item{inst}{(\link{FSelectInstanceBatchSingleCrit} | \link{FSelectInstanceBatchMultiCrit}).}
}
\description{
Most assertion functions ensure the right class attribute, and optionally additional properties.
}
\keyword{internal}
