% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_filters.R
\name{mlr_filters}
\alias{mlr_filters}
\title{Dictionary of Filters}
\format{\link{R6Class} object}
\description{
A simple \link{Dictionary} storing objects of class \link{Filter}.
Each Filter has an associated help page, see \code{mlr_filters_[id]}.
}
\section{Usage}{


See \link{Dictionary}.
}

\examples{
mlr_filters$keys()
as.data.table(mlr_filters)
mlr_filters$get("mim")
}
\seealso{
Other Filter: \code{\link{Filter}},
  \code{\link{mlr_filters_auc}},
  \code{\link{mlr_filters_cmim}},
  \code{\link{mlr_filters_disr}},
  \code{\link{mlr_filters_gain_ratio}},
  \code{\link{mlr_filters_information_gain}},
  \code{\link{mlr_filters_jmim}},
  \code{\link{mlr_filters_jmi}},
  \code{\link{mlr_filters_kruskal_test}},
  \code{\link{mlr_filters_linear_correlation}},
  \code{\link{mlr_filters_mim}},
  \code{\link{mlr_filters_njmim}},
  \code{\link{mlr_filters_rank_correlation}},
  \code{\link{mlr_filters_symmetrical_uncertainty}},
  \code{\link{mlr_filters_variable_importance}},
  \code{\link{mlr_filters_variance}}
}
\concept{Dictionary}
\concept{Filter}
