% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterDISR.R
\docType{data}
\name{mlr_filters_disr}
\alias{mlr_filters_disr}
\alias{FilterDISR}
\title{Double Input Symmetrical Relevance Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterDISR
}
\description{
Double input symmetrical relevance filter.
Calls \code{\link[praznik:DISR]{praznik::DISR()}}.
}
\examples{
task = mlr3::mlr_tasks$get("iris")
filter = FilterDISR$new()
filter$calculate(task)
head(as.data.table(filter), 3)
}
\seealso{
Other Filter: \code{\link{Filter}},
  \code{\link{mlr_filters_auc}},
  \code{\link{mlr_filters_cmim}},
  \code{\link{mlr_filters_gain_ratio}},
  \code{\link{mlr_filters_information_gain}},
  \code{\link{mlr_filters_jmim}},
  \code{\link{mlr_filters_jmi}},
  \code{\link{mlr_filters_kruskal_test}},
  \code{\link{mlr_filters_linear_correlation}},
  \code{\link{mlr_filters_mim}},
  \code{\link{mlr_filters_njmim}},
  \code{\link{mlr_filters_rank_correlation}},
  \code{\link{mlr_filters_symmetrical_uncertainty}},
  \code{\link{mlr_filters_variance}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
