% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_fselectors.R
\docType{data}
\name{mlr_fselectors}
\alias{mlr_fselectors}
\title{mlr_fselectors}
\format{
An object of class \code{DictionaryFSelect} (inherits from \code{Dictionary}, \code{R6}) of length 13.
}
\usage{
mlr_fselectors
}
\description{
A \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} storing objects of class \link{FSelector}.
}
\keyword{datasets}
