% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSelectorAsyncDesignPoints.R
\name{mlr_fselectors_async_design_points}
\alias{mlr_fselectors_async_design_points}
\alias{FSelectorAsyncDesignPoints}
\title{Feature Selection with Asynchronous Design Points}
\description{
Subclass for asynchronous design points feature selection.
}
\section{Dictionary}{

This \link{FSelector} can be instantiated with the associated sugar function \code{\link[=fs]{fs()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{fs("async_design_points")
}\if{html}{\out{</div>}}
}

\section{Parameters}{


\describe{
\item{\code{design}}{\link[data.table:data.table]{data.table::data.table}\cr
Design points to try in search, one per row.}
}

}

\seealso{
Other FSelectorAsync:
\code{\link{mlr_fselectors_async_exhaustive_search}},
\code{\link{mlr_fselectors_async_random_search}}
}
\concept{FSelectorAsync}
\section{Super classes}{
\code{\link[mlr3fselect:FSelector]{FSelector}} -> \code{\link[mlr3fselect:FSelectorAsync]{FSelectorAsync}} -> \code{\link[mlr3fselect:FSelectorAsyncFromOptimizerAsync]{FSelectorAsyncFromOptimizerAsync}} -> \code{FSelectorAsyncDesignPoints}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-FSelectorAsyncDesignPoints-initialize}{\code{FSelectorAsyncDesignPoints$new()}}
    \item \href{#method-FSelectorAsyncDesignPoints-clone}{\code{FSelectorAsyncDesignPoints$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelector" data-id="format"><a href='../../mlr3fselect/html/FSelector.html#method-FSelector-format'><code>FSelector$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelector" data-id="help"><a href='../../mlr3fselect/html/FSelector.html#method-FSelector-help'><code>FSelector$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelector" data-id="print"><a href='../../mlr3fselect/html/FSelector.html#method-FSelector-print'><code>FSelector$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelectorAsyncFromOptimizerAsync" data-id="optimize"><a href='../../mlr3fselect/html/FSelectorAsyncFromOptimizerAsync.html#method-FSelectorAsyncFromOptimizerAsync-optimize'><code>FSelectorAsyncFromOptimizerAsync$optimize()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FSelectorAsyncDesignPoints-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-FSelectorAsyncDesignPoints-initialize}{}}}
\subsection{\code{FSelectorAsyncDesignPoints$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{FSelectorAsyncDesignPoints$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FSelectorAsyncDesignPoints-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FSelectorAsyncDesignPoints-clone}{}}}
\subsection{\code{FSelectorAsyncDesignPoints$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{FSelectorAsyncDesignPoints$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
