% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSelectorShadowVariableSearch.R
\name{mlr_fselectors_shadow_variable_search}
\alias{mlr_fselectors_shadow_variable_search}
\alias{FSelectorShadowVariableSearch}
\title{Feature Selection via Shadow Variable Search}
\source{
Thomas J, Hepp T, Mayr A, Bischl B (2017).
\dQuote{Probing for Sparse and Fast Variable Selection with Model-Based Boosting.}
\emph{Computational and Mathematical Methods in Medicine}, \bold{2017}, 1--8.
\doi{10.1155/2017/1421409}.

Wu Y, Boos DD, Stefanski LA (2007).
\dQuote{Controlling Variable Selection by the Addition of Pseudovariables.}
\emph{Journal of the American Statistical Association}, \bold{102}(477), 235--243.
\doi{10.1198/016214506000000843}.
}
\description{
Shadow variable search creates for each feature a permutated copy and stops when one of them is selected.

The feature selection terminates itself when the first shadow variable is selected.
It is not necessary to set a termination criterion.
}
\section{Dictionary}{

This \link{FSelector} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary}
\link{mlr_fselectors} or with the associated sugar function \code{\link[=fs]{fs()}}:\preformatted{mlr_fselectors$get("shadow_variable_search")
fs("shadow_variable_search")
}
}

\examples{
# retrieve task
task = tsk("pima")

# load learner
learner = lrn("classif.rpart")

\donttest{
# feature selection on the pima indians diabetes data set
instance = fselect(
  method = "shadow_variable_search",
  task = task,
  learner = learner,
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
)

# best performing feature subset
instance$result

# all evaluated feature subsets
as.data.table(instance$archive)

# subset the task and fit the final model
task$select(instance$result_feature_set)
learner$train(task)
}
}
\section{Super class}{
\code{\link[mlr3fselect:FSelector]{mlr3fselect::FSelector}} -> \code{FSelectorShadowVariableSearch}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FSelectorShadowVariableSearch$new()}}
\item \href{#method-optimization_path}{\code{FSelectorShadowVariableSearch$optimization_path()}}
\item \href{#method-clone}{\code{FSelectorShadowVariableSearch$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelector" data-id="format">}\href{../../mlr3fselect/html/FSelector.html#method-format}{\code{mlr3fselect::FSelector$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelector" data-id="help">}\href{../../mlr3fselect/html/FSelector.html#method-help}{\code{mlr3fselect::FSelector$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelector" data-id="optimize">}\href{../../mlr3fselect/html/FSelector.html#method-optimize}{\code{mlr3fselect::FSelector$optimize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelector" data-id="print">}\href{../../mlr3fselect/html/FSelector.html#method-print}{\code{mlr3fselect::FSelector$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectorShadowVariableSearch$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimization_path"></a>}}
\if{latex}{\out{\hypertarget{method-optimization_path}{}}}
\subsection{Method \code{optimization_path()}}{
Returns the optimization path.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectorShadowVariableSearch$optimization_path(inst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inst}}{(\link{FSelectInstanceSingleCrit})\cr
Instance optimized with \link{FSelectorShadowVariableSearch}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link[data.table:data.table]{data.table::data.table}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectorShadowVariableSearch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
