% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskClassif_titanic.R
\name{mlr_tasks_titanic}
\alias{mlr_tasks_titanic}
\title{Titanic Classification Task}
\description{
Titanic data set from the opendatasoft repository
(\url{https://public.opendatasoft.com/explore/dataset/titanic-passengers}).
}
\section{Construction}{
\preformatted{mlr_tasks$get("titanic")
tsk("titanic")
}
}

