% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_terminators.R
\docType{data}
\name{mlr_terminators}
\alias{mlr_terminators}
\title{Dictionary of Terminators}
\format{An object of class \code{DictionaryTerminator} (inherits from \code{Dictionary}, \code{R6}) of length 13.}
\usage{
mlr_terminators
}
\description{
A simple \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} storing objects of class \link{Terminator}.
Each terminator has an associated help page, see \code{mlr_terminators_[id]}.

This dictionary can get populated with additional terminators by add-on packages.

For a more convenient way to retrieve and construct terminator, see \code{\link[=term]{term()}}.
}
\examples{
term("evals", n_evals = 10)
}
\seealso{
Sugar function: \code{\link[=term]{term()}}
}
\keyword{datasets}
