% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerEvolutionary.R
\name{OptimizerEvolutionary}
\alias{OptimizerEvolutionary}
\title{Feature Selection via Evolutionary Search}
\description{
Subclass for evolutionary feature selection. Calls \code{\link[ecr:ecr]{ecr::ecr()}} from package
\CRANpkg{ecr}.
}
\section{Dictionary}{

This \link[bbotk:Optimizer]{bbotk::Optimizer} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary}
\link{mlr_optimizers} or with the associated sugar function \code{\link[=opt]{opt()}}:\preformatted{mlr_optimizers$get("evolutionary")
opt("evolutionary")
}
}

\section{Parameters}{

\describe{
\item{\code{mu}}{\code{integer(1)}}
\item{\code{lambda}}{\code{integer(1)}}
\item{\code{p.recomb}}{\code{double(1)}}
\item{\code{p.mut}}{\code{double(1)}}
\item{\code{survival.strategy}}{\code{character(1)}}
\item{\code{n.elite}}{\code{integer(1)}}
\item{\code{initial.solutions}}{list of \code{integer()}}
\item{\code{parent.selector}}{\code{ecr} selector}
\item{\code{survival.selector}}{\code{ecr} selector}
}

For the meaning of the control parameter, see \code{\link[ecr:ecr]{ecr::ecr()}}. Note that \code{mu}
and \code{lambda} must be set by the user. The \code{terminators} parameter is replaced
by the \link{Terminator} subclasses.
}

\examples{
library(mlr3)

terminator = term("evals", n_evals = 10)
instance = FSelectInstance$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measure = msr("classif.ce"),
  terminator = terminator
)

fs = opt("evolutionary", mu = 10, lambda = 5)
fs$optimize(instance)
instance$result
instance$archive$data
}
\section{Super class}{
\code{\link[bbotk:Optimizer]{bbotk::Optimizer}} -> \code{OptimizerEvolutionary}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OptimizerEvolutionary$new()}}
\item \href{#method-clone}{\code{OptimizerEvolutionary$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="format">}\href{../../bbotk/html/Optimizer.html#method-format}{\code{bbotk::Optimizer$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="optimize">}\href{../../bbotk/html/Optimizer.html#method-optimize}{\code{bbotk::Optimizer$optimize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="print">}\href{../../bbotk/html/Optimizer.html#method-print}{\code{bbotk::Optimizer$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerEvolutionary$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerEvolutionary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
