% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceEvaluator.R
\name{PerformanceEvaluator}
\alias{PerformanceEvaluator}
\title{Abstract PerformanceEvaluator Class}
\description{
\code{PerformanceEvaluator} class that implements the performance evaluation on a set of feature combinations. A pe is an object that stores all informations that are necesarry to conduct a feature selection (\code{mlr3::Task}, \code{mlr3::Learner}, \code{mlr3::Resampling}).
}
\section{Usage}{
\preformatted{# Construction
pe = PerformanceEvaluator$new()

# Public members
pe$task
pe$learner
pe$resampling
pe$bmr

# Public methods
pe$eval_states(states)
pe$get_best()
}
}

\section{Arguments}{

\itemize{
\item \code{task} (\code{mlr3::Task}):
The task that we want to evaluate.
\item \code{learner} (\code{mlr3::Learner}):
The learner that we want to evaluate.
\item \code{resampling} (\code{mlr3::Resampling}):
The Resampling method that is used to evaluate the learner.
}
}

\section{Details}{

\itemize{
\item \code{$new()} creates a new object of class \link{PerformanceEvaluator}.
\item \code{$task} (\code{mlr3::Task}) the task for which the feature selection should be conducted.
\item \code{$learner} (\code{mlr3::Learner}) the algorithm for which the feature selection should be conducted.
\item \code{$resampling} (\code{mlr3::Resampling}) strategy to evaluate a feature combination
\item \code{$bmr} (\code{list}) of (\code{mlr3::BenchmarkResult}) objects. Each entry corresponds to one batch or step depending one the used feature selection method.
\item \code{eval_states(states)} evaluates the feature combinations \code{states}
\item \code{get_best()} returns selected features with the best performance of each entry in \code{$bmr}
}
}

\examples{
task = mlr3::mlr_tasks$get("iris")
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("holdout")
pe = PerformanceEvaluator$new(task, learner, resampling)
}
\concept{PerformanceEvaluator}
\keyword{internal}
