% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{Terminator}
\alias{Terminator}
\title{Re-export of \code{Terminator}

See \link[bbotk:Terminator]{bbotk::Terminator}.}
\description{
Re-export of \code{Terminator}

See \link[bbotk:Terminator]{bbotk::Terminator}.
}
