% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorClockTime.R
\name{TerminatorClockTime}
\alias{TerminatorClockTime}
\alias{mlr_terminators_clock_time}
\title{Terminator that stops according to the clock time}
\description{
Class to terminate the optimization either after the complete process took a number of seconds on the clock
or a fixed time point has been reached (as reported by \code{\link[=Sys.time]{Sys.time()}}).
}
\section{Parameters}{

\itemize{
\item \code{secs} :: \code{numeric(1)}\cr
Maximum allowed time, in seconds, default is 100.
Mutually exclusive with argument \code{stop_time}.
\item \code{stop_time} :: \code{POSIXct(1)}\cr
Terminator stops after this point in time.
Mutually exclusive with argument \code{secs}.
}
}

\examples{
term("clock_time", secs = 1800)

stop_time = as.POSIXct("2030-01-01 00:00:00")
term("clock_time", stop_time = stop_time)
}
\section{Super class}{
\code{\link[mlr3fselect:Terminator]{mlr3fselect::Terminator}} -> \code{TerminatorClockTime}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TerminatorClockTime$new()}}
\item \href{#method-is_terminated}{\code{TerminatorClockTime$is_terminated()}}
\item \href{#method-clone}{\code{TerminatorClockTime$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="Terminator" data-id="format">}\href{../../mlr3fselect/html/Terminator.html#method-format}{\code{mlr3fselect::Terminator$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="Terminator" data-id="print">}\href{../../mlr3fselect/html/Terminator.html#method-print}{\code{mlr3fselect::Terminator$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create new \code{TerminatorClockTime} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorClockTime$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{TerminatorClockTime} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_terminated"></a>}}
\subsection{Method \code{is_terminated()}}{
Is \code{TRUE} if after the complete process took a number of seconds on the clock
or a fixed time point has been reached, and \code{FALSE} otherwise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorClockTime$is_terminated(instance)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{object of class \link{FSelect}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{logical(1)}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorClockTime$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
