% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{TerminatorEvals}
\alias{TerminatorEvals}
\title{Re-export of \code{TerminatorEvals}

See \link[bbotk:TerminatorEvals]{bbotk::TerminatorEvals}.}
\description{
Re-export of \code{TerminatorEvals}

See \link[bbotk:TerminatorEvals]{bbotk::TerminatorEvals}.
}
