% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorEvaluations.R
\docType{data}
\name{TerminatorEvaluations}
\alias{TerminatorEvaluations}
\title{TerminatorEvaluations}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Terminator}.}
\usage{
TerminatorEvaluations
}
\description{
Feature selection terminates depending on the number of evaluations. An evaluation is defined by one resampling of a feature set.
}
\section{Construction}{
\preformatted{fs = TerminatorEvaluations$new(max_evaluations)
}
\itemize{
\item \code{max_evaluations} (\code{integer(1)})\cr Maximum number of evaluations.
}
}

\section{Fields}{

See \link{Terminator}.
}

\section{Methods}{

See \link{Terminator}.
}

\examples{
tm = TerminatorEvaluations$new(max_evaluations = 100)
}
\seealso{
Other Terminator: \code{\link{TerminatorPerformanceStep}},
  \code{\link{TerminatorRuntime}}, \code{\link{Terminator}}
}
\concept{Terminator}
\keyword{datasets}
