% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorModelTime.R
\name{TerminatorModelTime}
\alias{TerminatorModelTime}
\alias{mlr_terminators_model_time}
\title{Terminator that stops after a budget of model evaluation time is depleted}
\description{
Class to terminate the optimization after a given model evaluation budget is exceeded.
The terminator measures the used time to train and predict all models contained
in the archive.
}
\section{Parameters}{

\itemize{
\item \code{secs} :: \code{numeric(1)}\cr
Maximum allowed time, in seconds, default is 0.
}
}

\examples{
TerminatorModelTime$new()
term("model_time", secs = 10 * 3600)
}
\section{Super class}{
\code{\link[mlr3fselect:Terminator]{mlr3fselect::Terminator}} -> \code{TerminatorModelTime}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TerminatorModelTime$new()}}
\item \href{#method-is_terminated}{\code{TerminatorModelTime$is_terminated()}}
\item \href{#method-clone}{\code{TerminatorModelTime$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="Terminator" data-id="format">}\href{../../mlr3fselect/html/Terminator.html#method-format}{\code{mlr3fselect::Terminator$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="Terminator" data-id="print">}\href{../../mlr3fselect/html/Terminator.html#method-print}{\code{mlr3fselect::Terminator$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create new \code{TerminatorModelTime} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorModelTime$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{TerminatorModelTime} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_terminated"></a>}}
\subsection{Method \code{is_terminated()}}{
Is \code{TRUE} if after a given model evaluation budget is exceeded, and \code{FALSE} otherwise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorModelTime$is_terminated(instance)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{object of class \link{FSelect}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{logical(1)}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorModelTime$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
