% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{TerminatorNone}
\alias{TerminatorNone}
\title{Re-export of \code{TerminatorNone}

See \link[bbotk:TerminatorNone]{bbotk::TerminatorNone}.}
\description{
Re-export of \code{TerminatorNone}

See \link[bbotk:TerminatorNone]{bbotk::TerminatorNone}.
}
