% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorPerformance.R
\docType{data}
\name{TerminatorPerformance}
\alias{TerminatorPerformance}
\title{TerminatorPerformance}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Terminator}.}
\usage{
TerminatorPerformance
}
\description{
Feature selection terminates depending on the performance.
}
\section{Construction}{
\preformatted{tm = TerminatorPerformance$new(threshold)
}
\itemize{
\item \code{threshold} (`numeric(1)``)\cr
}
}

\section{Fields}{

See \link{Terminator}.
}

\section{Methods}{

See \link{Terminator}.
}

\examples{
tm = TerminatorPerformance$new(threshold = 0.76)
}
\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorPerformanceStep}},
  \code{\link{TerminatorRuntime}}, \code{\link{Terminator}}
}
\concept{Terminator}
\keyword{datasets}
