% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorPerfReached.R
\name{TerminatorPerfReached}
\alias{TerminatorPerfReached}
\alias{mlr_terminators_perf_reached}
\title{Terminator that stops when a performance level has been reached}
\description{
Class to terminate the tuning after a performance level has been hit.
}
\section{Parameters}{

\itemize{
\item \code{level} :: \code{numeric(1)}\cr
Performance level that needs to be reached, default is 0.
Terminates if the performance exceeds (respective measure has to be maximized) or
falls below (respective measure has to be minimized) this value.
}
}

\examples{
TerminatorPerfReached$new()
term("perf_reached")
}
\section{Super class}{
\code{\link[mlr3fselect:Terminator]{mlr3fselect::Terminator}} -> \code{TerminatorPerfReached}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TerminatorPerfReached$new()}}
\item \href{#method-is_terminated}{\code{TerminatorPerfReached$is_terminated()}}
\item \href{#method-clone}{\code{TerminatorPerfReached$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="Terminator" data-id="format">}\href{../../mlr3fselect/html/Terminator.html#method-format}{\code{mlr3fselect::Terminator$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="Terminator" data-id="print">}\href{../../mlr3fselect/html/Terminator.html#method-print}{\code{mlr3fselect::Terminator$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create new \code{TerminatorPerfReached} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorPerfReached$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{TerminatorPerfReached} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_terminated"></a>}}
\subsection{Method \code{is_terminated()}}{
Is \code{TRUE} after a performance level has been hit, and \code{FALSE} otherwise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorPerfReached$is_terminated(instance)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{object of class \link{FSelect}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{logical(1)}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorPerfReached$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
