% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{TerminatorRunTime}
\alias{TerminatorRunTime}
\title{Re-export of \code{TerminatorRunTime}

See \link[bbotk:TerminatorRunTime]{bbotk::TerminatorRunTime}.}
\description{
Re-export of \code{TerminatorRunTime}

See \link[bbotk:TerminatorRunTime]{bbotk::TerminatorRunTime}.
}
