% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{TerminatorStagnation}
\alias{TerminatorStagnation}
\title{Re-export of \code{TerminatorStagnation}

See \link[bbotk:TerminatorStagnation]{bbotk::TerminatorStagnation}.}
\description{
Re-export of \code{TerminatorStagnation}

See \link[bbotk:TerminatorStagnation]{bbotk::TerminatorStagnation}.
}
