% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFSelect.R
\name{AutoFSelect}
\alias{AutoFSelect}
\title{AutoFeselect}
\description{
The \code{AutoFSelect} is a \link[mlr3:Learner]{mlr3::Learner} which automatically selects the optimal feature set and
fits the model on the training data with the optimal feature set.
}
\examples{
library(mlr3)

task = tsk("iris")
learner = lrn("classif.rpart")
resampling = rsmp("holdout")
measures = msr("classif.ce")

terminator = term("evals", n_evals = 15)
fs = fs("exhaustive")
afs = AutoFSelect$new(learner, resampling, measures, terminator, fs)
afs$store_fselect_instance = TRUE

afs$train(task)
afs$model
afs$learner
}
\section{Super class}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{AutoFSelect}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{instance_args}}{\code{list()}}

\item{\code{fselect}}{\link{FSelect}}

\item{\code{store_fselect_instance}}{\code{logical(1)}}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{learner}}{\link[mlr3:Learner]{mlr3::Learner}}

\item{\code{fselect_instance}}{\link{FSelectInstance}}

\item{\code{fselect_result}}{\code{list()} Result of the feature selection i.e. the optimal feature set and its estimated performances.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AutoFSelect$new()}}
\item \href{#method-train_internal}{\code{AutoFSelect$train_internal()}}
\item \href{#method-predict_internal}{\code{AutoFSelect$predict_internal()}}
\item \href{#method-archive}{\code{AutoFSelect$archive()}}
\item \href{#method-clone}{\code{AutoFSelect$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create new \code{AutoFSelect} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoFSelect$new(
  learner,
  resampling,
  measures,
  terminator,
  fselect,
  bm_args = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learner}}{\link[mlr3:Learner]{mlr3::Learner}}

\item{\code{resampling}}{\link[mlr3:Resampling]{mlr3::Resampling}}

\item{\code{measures}}{list of \link[mlr3:Measure]{mlr3::Measure}}

\item{\code{terminator}}{\link{Terminator}}

\item{\code{fselect}}{\link{FSelect}}

\item{\code{bm_args}}{named \code{list()} Further arguments for \code{\link[mlr3:benchmark]{mlr3::benchmark()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{AutoFSelect}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train_internal"></a>}}
\subsection{Method \code{train_internal()}}{
Selects the optimal feature set, applies the feature set to the task
and fits the model on the training data.
If \code{store_fselect_instance} is \code{TRUE}, the \link[mlr3:Learner]{mlr3::Learner} with the \link{FSelectInstance}
is returned, otherwise just the \link[mlr3:Learner]{mlr3::Learner}. For internal use.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoFSelect$train_internal(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{\link[mlr3:Task]{mlr3::Task}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list(\link[mlr3:Learner]{mlr3::Learner}) or list(\link[mlr3:Learner]{mlr3::Learner}, \link{FSelectInstance})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_internal"></a>}}
\subsection{Method \code{predict_internal()}}{
Creates a new \link[mlr3:Prediction]{mlr3::Prediction} based on the learner fitted
on the training data with the optimal feature subset.
For internal use.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoFSelect$predict_internal(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{\link[mlr3:Task]{mlr3::Task}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link[mlr3:Prediction]{mlr3::Prediction}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-archive"></a>}}
\subsection{Method \code{archive()}}{
Returns a table of contained resample results with corresponding feature sets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoFSelect$archive()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\link[data.table:data.table]{data.table::data.table}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoFSelect$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
