% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_fselect.R
\name{ensemble_fselect}
\alias{ensemble_fselect}
\title{Ensemble Feature Selection}
\usage{
ensemble_fselect(
  fselector,
  task,
  learners,
  outer_resampling,
  inner_resampling,
  measure,
  terminator
)
}
\arguments{
\item{fselector}{(\link{FSelector})\cr
Optimization algorithm.}

\item{task}{(\link[mlr3:Task]{mlr3::Task})\cr
Task to operate on.}

\item{learners}{(list of \link[mlr3:Learner]{mlr3::Learner})\cr
The learners to be used for feature selection.}

\item{outer_resampling}{(\link[mlr3:Resampling]{mlr3::Resampling})\cr
The outer resampling strategy.
The number of iterations must match the number of learners.}

\item{inner_resampling}{(\link[mlr3:Resampling]{mlr3::Resampling})\cr
The inner resampling strategy used by the \link{FSelector}.}

\item{measure}{(\link[mlr3:Measure]{mlr3::Measure})\cr
Measure to optimize. If \code{NULL}, default measure is used.}

\item{terminator}{(\link{Terminator})\cr
Stop criterion of the feature selection.}
}
\description{
Ensemble feature selection using multiple learners.
}
\examples{
\donttest{

  ensemble_fselect(
    fselector = fs("random_search"),
    task = tsk("sonar"),
    learners = lrns(c("classif.rpart", "classif.featureless")),
    outer_resampling = rsmp("subsampling", repeats = 2),
    inner_resampling = rsmp("cv", folds = 3),
    measure = msr("classif.ce"),
    terminator = trm("evals", n_evals = 10)
  )
}
}
