% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{FeatureSelection}
\alias{FeatureSelection}
\title{Abstract FeatureSelection Class}
\description{
\code{FeatureSelection} class that implements the main functionality each fs must have. A fs is an object that describes the optimization method for choosing the features given within the \code{[PerformanceEvaluator]} object.
}
\section{Usage}{
\preformatted{# Construction
fs = FeatureSelectionr$new(id, pe, tm, settings = list())

# public members
fs$id
fs$pe
fs$tm
fs$settings

# public methods
fs$calculate()
}
}

\section{Arguments}{

\itemize{
\item \code{id} (\code{character(1)}):\cr
The id of the FeatureSelection.
\item \code{pe} (\code{[PerformanceEvaluator]}).
\item \code{tm} (\code{[Terminator]}).
\item \code{settings} (\code{list}):\cr
The settings for the FeatureSelection.
}
}

\section{Details}{

\itemize{
\item \code{$new()} creates a new object of class \code{[FeatureSelection]}.
\item \code{$id} stores an identifier for this \code{[FeatureSelection]}.
\item \code{$pe} stores the \link{PerformanceEvaluator} to optimize.
\item \code{$tm} stores the \code{[Terminator]}.
\item \code{$settings} is a list of settings for this \code{[FeatureSelection]}.
\item \code{$state} stores currently evaluated 0/1 encoded feature combinations.
\item \code{$calculate()} performs the feature selection, until the budget of the \code{[Terminator]} in the \code{[PerformanceEvaluator]} is exhausted.
}
}

\seealso{
Other FeatureSelection: \code{\link{FeatureSelectionGenetic}},
  \code{\link{FeatureSelectionRandom}},
  \code{\link{FeatureSelectionSequential}}
}
\concept{FeatureSelection}
