% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelectionExhaustive.R
\name{FeatureSelectionExhaustive}
\alias{FeatureSelectionExhaustive}
\title{FeatureSelectionExhaustive}
\description{
FeatureSelection child class to conduct exhaustive search
}
\section{Usage}{
\preformatted{fs = FeatureSelectionExhaustive$new()
}

See \link{FeatureSelection} for a description of the interface.
}

\section{Arguments}{

\itemize{
\item \code{pe} (\code{[PerformanceEvaluator]}).
\item \code{tm} (\code{[Terminator]}).
\item \code{max_features} (\code{integer(1)})
Maximum number of features
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{FeatureSelectionExhaustive}.
\code{$get_result()} Returns best feature combination.
The interface is described in \link{FeatureSelection}.
}

\examples{
task = mlr3::mlr_tasks$get("pima")
task$select(c("age", "glucose", "insulin", "mass"))
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("cv", param_vals = list(folds = 5L))
pe = PerformanceEvaluator$new(task = task, learner = learner, resampling = resampling)
tm = TerminatorRuntime$new(max_time = 20, units = "secs")
fs = FeatureSelectionExhaustive$new(pe = pe, tm = tm, max_features = 3)
fs$calculate()
fs$get_result()
}
\seealso{
Other FeatureSelection: \code{\link{FeatureSelectionGenetic}},
  \code{\link{FeatureSelectionRandom}},
  \code{\link{FeatureSelectionSequential}},
  \code{\link{FeatureSelection}}
}
\concept{FeatureSelection}
