% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelectionRandom.R
\docType{data}
\name{FeatureSelectionRandom}
\alias{FeatureSelectionRandom}
\title{FeatureSelectionRandom}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{FeatureSelection}.}
\usage{
FeatureSelectionRandom
}
\description{
Random feature selection wrapper.
}
\section{Construction}{
\preformatted{fs = FeatureSelectionRandom$new(pe, tm, measure, param_vals)
}

For arguments, see \link{FeatureSelection}.
}

\section{Fields}{

See \link{FeatureSelection}.
}

\section{Methods}{

See \link{FeatureSelection} and additionally:
\itemize{
\item \code{$get_result()}\cr Returns best feature set.
}
}

\section{Parameter Values}{

See \link{FeatureSelection}.
}

\examples{
task = mlr3::mlr_tasks$get("pima")
measure = mlr3::mlr_measures$get(c("classif.acc"))
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("cv", param_vals = list(folds = 5L))
resampling$instantiate(task)
pe = PerformanceEvaluator$new(task, learner, resampling)
tm = TerminatorEvaluations$new(max_evaluations = 50)
fs = FeatureSelectionRandom$new(pe, tm, measure,
                                param_vals = list(max_features = 4))
fs$calculate()
fs$get_result()
}
\seealso{
Other FeatureSelection: \code{\link{FeatureSelectionExhaustive}},
  \code{\link{FeatureSelectionGenetic}},
  \code{\link{FeatureSelectionSequential}},
  \code{\link{FeatureSelection}}
}
\concept{FeatureSelection}
\keyword{datasets}
