% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterDISR.R
\docType{data}
\name{FilterDISR}
\alias{FilterDISR}
\alias{mlr_filters_disr}
\title{Double Input Symmetrical Relevance Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterDISR
}
\description{
Double input symmetrical relevance filter. Calls
\code{\link[praznik:DISR]{praznik::DISR()}}.
}
\details{
This filter supports partial scoring via hyperparameter \code{k}. To use
it, set \code{k} during construction via \code{param_vals.} By default all filter
scores are calculated and the default of \code{k = 3} in the ParamSet does not
apply.
}
\examples{
task = mlr3::mlr_tasks$get("iris")
filter = FilterDISR$new()
filter$calculate(task)
as.data.table(filter)[1:3]
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterCarScore}},
  \code{\link{FilterCorrelation}},
  \code{\link{FilterEmbedded}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}}, \code{\link{FilterMIM}},
  \code{\link{FilterMRMR}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
