% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterEmbedded.R
\docType{data}
\name{FilterEmbedded}
\alias{FilterEmbedded}
\alias{mlr_filters_variable_importance}
\title{Filter which uses learner integrated methods}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterEmbedded
}
\description{
Variable Importance filter using learner embedded methods. Takes
a \link[mlr3:Learner]{mlr3::Learner} which supports retrieving the variable importance
(property "importance"), fits the model and uses the importance values as
filter scores.
}
\examples{
task = mlr3::mlr_tasks$get("iris")
learner = mlr3::mlr_learners$get("classif.rpart")
filter = FilterEmbedded$new(learner = learner)
filter$calculate(task)
as.data.table(filter)[1:3]
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterCarScore}},
  \code{\link{FilterCorrelation}},
  \code{\link{FilterDISR}}, \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}}, \code{\link{FilterMIM}},
  \code{\link{FilterMRMR}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
