% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterInformationGain.R
\docType{data}
\name{FilterInformationGain}
\alias{FilterInformationGain}
\alias{mlr_filters_information_gain}
\title{Information Gain Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{FilterResult}.}
\usage{
FilterInformationGain
}
\description{
Information gain filter.
Calls \code{\link[FSelectorRcpp:information_gain]{FSelectorRcpp::information_gain()}}.
The target variable of regression tasks is automatically binned (argument \code{equal} in \code{\link[FSelectorRcpp:information_gain]{FSelectorRcpp::information_gain()}}).
}
\examples{
task = mlr3::mlr_tasks$get("sonar")
filter = FilterInformationGain$new()
filter$calculate(task)
as.data.table(filter)[1:3]
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterDISR}},
  \code{\link{FilterGainRatio}}, \code{\link{FilterJMIM}},
  \code{\link{FilterJMI}}, \code{\link{FilterKruskalTest}},
  \code{\link{FilterLinearCorrelation}},
  \code{\link{FilterMIM}}, \code{\link{FilterMRMR}},
  \code{\link{FilterNJMIM}},
  \code{\link{FilterRankCorrelation}},
  \code{\link{FilterResult}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
