% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterInformationGain.R
\name{FilterInformationGain}
\alias{FilterInformationGain}
\title{Gain Ratio}
\description{
FilterGain Ratio
}
\section{Usage}{
\preformatted{filter = FilterInformationGain$new()
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{FilterInformationGain}.


\itemize{
\item \code{$calculate(task)} calculates the filter values.
\item \code{$new()} creates a new object of class \link{Filter}.
\item \code{$filter()} filters a \link{Task} using specific criteria
\itemize{
\item returns a subsetted \link{Task}
}
\item \code{$id} stores an identifier for this \link{Filter}.
\item \code{$filter_values} stores the calculated filter values.
\item \code{$packages} stores the names of required packages.
\item \code{$settings} is a list of hyperparamter settings for this \link{Filter}.
}
}

\examples{
task = mlr_tasks$get("sonar")
filter = FilterInformationGain$new()
filter$calculate(task)
head(as.data.table(filter), 3)
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterDISR}},
  \code{\link{FilterGainRatio}}, \code{\link{FilterJMIM}},
  \code{\link{FilterJMI}}, \code{\link{FilterKruskalTest}},
  \code{\link{FilterLinearCorrelation}},
  \code{\link{FilterMIM}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterRankCorrelation}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{Filter}}
}
\concept{Filter}
