% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterJMI.R
\docType{data}
\name{FilterJMI}
\alias{FilterJMI}
\alias{mlr_filters_jmi}
\title{Joint Mutual Information Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterJMI
}
\description{
Joint mutual information filter.
Calls \code{\link[praznik:JMI]{praznik::JMI()}}.
}
\examples{
task = mlr3::mlr_tasks$get("iris")
filter = FilterJMI$new()
filter$calculate(task)
as.data.table(filter)[1:3]
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterCarScore}},
  \code{\link{FilterDISR}}, \code{\link{FilterEmbedded}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}},
  \code{\link{FilterKruskalTest}},
  \code{\link{FilterLinearCorrelation}},
  \code{\link{FilterMIM}}, \code{\link{FilterMRMR}},
  \code{\link{FilterNJMIM}},
  \code{\link{FilterRankCorrelation}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
