% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterKruskalTest.R
\name{FilterKruskalTest}
\alias{FilterKruskalTest}
\title{Kruskal Test}
\description{
FilterKruskalTest
}
\section{Usage}{
\preformatted{filter = FilterKruskalTest$new()
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{FilterKruskalTest}.


\itemize{
\item \code{$calculate(task)} calculates the filter values.
\item \code{$new()} creates a new object of class \link{Filter}.
\item \code{$filter()} filters a \link{Task} using specific criteria
\itemize{
\item returns a subsetted \link{Task}
}
\item \code{$id} stores an identifier for this \link{Filter}.
\item \code{$filter_values} stores the calculated filter values.
\item \code{$packages} stores the names of required packages.
\item \code{$settings} is a list of hyperparamter settings for this \link{Filter}.
}
}

\examples{
task = mlr3::mlr_tasks$get("iris")
filter = FilterKruskalTest$new()
filter$calculate(task)
filter$filter_abs(task, 2)
head(as.data.table(filter), 3)
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterDISR}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterLinearCorrelation}},
  \code{\link{FilterMIM}}, \code{\link{FilterNJMIM}},
  \code{\link{FilterRankCorrelation}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
