% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterLinearCorrelation.R
\docType{data}
\name{FilterLinearCorrelation}
\alias{FilterLinearCorrelation}
\alias{mlr_filters_linear_correlation}
\title{Linear Correlation Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterLinearCorrelation
}
\description{
Linear correlation filter.
Calls \code{\link[stats:cor]{stats::cor()}}.
}
\examples{
task = mlr3::mlr_tasks$get("mtcars")
filter = FilterLinearCorrelation$new()
filter$calculate(task)
head(as.data.table(filter), 3)
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterDISR}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}}, \code{\link{FilterMIM}},
  \code{\link{FilterNJMIM}},
  \code{\link{FilterRankCorrelation}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariableImportance}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
