% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterMRMR.R
\docType{data}
\name{FilterMRMR}
\alias{FilterMRMR}
\alias{mlr_filters_MRMR}
\title{Minimum redundancy maximal relevancy filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterMRMR
}
\description{
Minimum redundancy maximal relevancy filter. Calls
\code{\link[praznik:MRMR]{praznik::MRMR()}}.
}
\details{
This filter supports partial scoring via hyperparameter \code{k}. To use
it, set \code{k} during construction via \code{param_vals.} By default all filter
scores are calculated and the default of \code{k = 3} in the ParamSet does not
apply.
}
\examples{
task = mlr3::mlr_tasks$get("iris")
filter = FilterMRMR$new()
filter$calculate(task)
head(as.data.table(filter), 3)
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterCarScore}},
  \code{\link{FilterCorrelation}},
  \code{\link{FilterDISR}}, \code{\link{FilterEmbedded}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}}, \code{\link{FilterMIM}},
  \code{\link{FilterNJMIM}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
