% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterNJMIM.R
\docType{data}
\name{FilterNJMIM}
\alias{FilterNJMIM}
\alias{mlr_filters_njmim}
\title{Minimal Normalised Joint Mutual Information Maximisation Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterNJMIM
}
\description{
Minimal normalised joint mutual information maximisation filter.
Calls \code{\link[praznik:NJMIM]{praznik::NJMIM()}}.
}
\examples{
task = mlr3::mlr_tasks$get("iris")
filter = FilterNJMIM$new()
filter$calculate(task)
head(as.data.table(filter), 3)
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterDISR}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}},
  \code{\link{FilterLinearCorrelation}},
  \code{\link{FilterMIM}},
  \code{\link{FilterRankCorrelation}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariableImportance}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
