% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterSymmetricalUncertainty.R
\docType{data}
\name{FilterSymmetricalUncertainty}
\alias{FilterSymmetricalUncertainty}
\alias{mlr_filters_symmetrical_uncertainty}
\title{Symmetrical Uncertainty Filter}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Filter}.}
\usage{
FilterSymmetricalUncertainty
}
\description{
Symmetrical uncertainty filter.
Calls \code{\link[FSelectorRcpp:information_gain]{FSelectorRcpp::information_gain()}}.
The target variable of regression tasks is automatically binned (argument \code{equal} in \code{\link[FSelectorRcpp:information_gain]{FSelectorRcpp::information_gain()}}).
}
\examples{
task = mlr3::mlr_tasks$get("sonar")
filter = FilterSymmetricalUncertainty$new()
filter$calculate(task)
head(as.data.table(filter), 3)
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterDISR}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}},
  \code{\link{FilterLinearCorrelation}},
  \code{\link{FilterMIM}}, \code{\link{FilterMRMR}},
  \code{\link{FilterNJMIM}},
  \code{\link{FilterRankCorrelation}},
  \code{\link{FilterVariableImportance}},
  \code{\link{FilterVariance}}, \code{\link{Filter}},
  \code{\link{mlr_filters}}
}
\concept{Filter}
\keyword{datasets}
