% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSelect.R
\name{FSelect}
\alias{FSelect}
\title{FSelect}
\description{
Abstract \code{FSelect} class that implements the base functionality each
\code{FSelect} subclass must provide. A \code{FSelect} object describes the feature
selection strategy, i.e. how to optimize the black-box function and its
feasible set defined by the \link{ObjectiveFSelect} object.

A list of measures can be passed to the instance, and they will always be all
evaluated. However, single-criteria algorithms optimize only the first
measure

A \code{FSelect} object must provide a private method \verb{$.assign_result(inst)} that writes the result to the \link{FSelectInstance} using the \verb{$assign_result()} method of
the instance in order to store the best selected feature
subset and its estimated performance vector at the end of the optimization.
}
\examples{
library(mlr3)

terminator = term("evals", n_evals = 10)
instance = FSelectInstance$new(
  task = tsk("iris"),
  learner = lrn("classif.rpart"),
  resampling = rsmp("holdout"),
  measures = msr("classif.ce"),
  terminator = terminator
)

fs = fs("random")
fs$optimize(instance)
instance$result
instance$archive$data
}
\section{Super class}{
\code{\link[bbotk:Optimizer]{bbotk::Optimizer}} -> \code{FSelect}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FSelect$new()}}
\item \href{#method-optimize}{\code{FSelect$optimize()}}
\item \href{#method-clone}{\code{FSelect$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="format">}\href{../../bbotk/html/Optimizer.html#method-format}{\code{bbotk::Optimizer$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="print">}\href{../../bbotk/html/Optimizer.html#method-print}{\code{bbotk::Optimizer$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelect$new(param_set, properties, packages = character(0))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param_set}}{\link[paradox:ParamSet]{paradox::ParamSet}}

\item{\code{properties}}{\code{character()}}

\item{\code{packages}}{\code{character()}}

\item{\code{param_classes}}{\code{character()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-optimize}{}}}
\subsection{Method \code{optimize()}}{
Performs the feature selection on a \link{FSelectInstance} until termination.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelect$optimize(inst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inst}}{\link{FSelectInstance}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelect$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
