% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSelect.R
\name{FSelect}
\alias{FSelect}
\title{FSelect}
\description{
Abstract \code{FSelect} class that implements the base functionality each \verb{FSelect*} class must provide.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{param_set}}{\link[paradox:ParamSet]{paradox::ParamSet}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FSelect$new()}}
\item \href{#method-select}{\code{FSelect$select()}}
\item \href{#method-clone}{\code{FSelect$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create new \code{FSelect} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelect$new(param_set)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param_set}}{\link[paradox:ParamSet]{paradox::ParamSet}
Set of control parameter for the feature selection.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{FSelect} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-select"></a>}}
\subsection{Method \code{select()}}{
Performce the feature selection on a \link{FSelectInstance} until termination.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelect$select(instance)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{\link{FSelectInstance}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelect$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
