% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSelectInstance.R
\name{FSelectInstance}
\alias{FSelectInstance}
\title{FSelectInstance Class}
\description{
Specifies a general feature selection scenario.

Evaluations of feature sets are performed in batches. Before and after a
batch is evaluated, the \link{Terminator} is queried for the remaining budget. If
the available budget is exhausted, an exception is raised, and no further
evaluations can be performed from this point on.

A list of measures can be passed to the instance, and they will always be all
evaluated. However, single-criteria tuners optimize only the first measure.

The \link{FSelect} object is also supposed to store its final result, consisting
of a selected feature set and associated estimated performance values, by
calling the method \code{instance$assign_result()}.
}
\examples{
library(mlr3)

# Objects required to define the performance evaluator
task = tsk("iris")
measures = msrs(c("classif.ce"))
learner = lrn("classif.rpart")
resampling = rsmp("cv")
terminator = term("evals", n_evals = 8)

instance = FSelectInstance$new(
  task = task,
  learner = learner,
  resampling = resampling,
  measures = measures,
  terminator = terminator
)

# Evaluate 4 feature sets
m = diag(4)
instance$eval_batch(m)
instance$archive()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task}}{(\link[mlr3:Task]{mlr3::Task}).}

\item{\code{learner}}{(\link[mlr3:Learner]{mlr3::Learner}).}

\item{\code{resampling}}{(\link[mlr3:Resampling]{mlr3::Resampling})}

\item{\code{measures}}{(list of \link[mlr3:Measure]{mlr3::Measure}).}

\item{\code{terminator}}{(\link{Terminator}).}

\item{\code{evaluator}}{(\link[bbotk:Evaluator]{bbotk::Evaluator}).}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{result}}{Result of the feature selection i.e. the optimal feature
set and its estimated performances.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FSelectInstance$new()}}
\item \href{#method-print}{\code{FSelectInstance$print()}}
\item \href{#method-assign_result}{\code{FSelectInstance$assign_result()}}
\item \href{#method-clone}{\code{FSelectInstance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$new(task, learner, resampling, measures, terminator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{\link[mlr3:Task]{mlr3::Task}}

\item{\code{learner}}{\link[mlr3:Learner]{mlr3::Learner}}

\item{\code{resampling}}{\link[mlr3:Resampling]{mlr3::Resampling}
Note that uninstantiated resamplings are instantiated during construction
so that all configurations.}

\item{\code{measures}}{list of \link[mlr3:Measure]{mlr3::Measure}}

\item{\code{terminator}}{\link{Terminator}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assign_result"></a>}}
\if{latex}{\out{\hypertarget{method-assign_result}{}}}
\subsection{Method \code{assign_result()}}{
The \link{FSelect} object writes the best found feature set
and estimated performance values here. For internal use.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$assign_result(feat, perf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feat}}{\code{character}
Must be character vector of feature names existing in \code{task}}

\item{\code{perf}}{\code{numeric}
Must be named numeric of performance measures, named with performance
IDs, regarding all elements in \code{measures}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
