% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSelectInstance.R
\name{FSelectInstance}
\alias{FSelectInstance}
\title{FSelectInstance Class}
\description{
Specifies a general feature selection scenario.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task}}{\link[mlr3:Task]{mlr3::Task}}

\item{\code{learner}}{\link[mlr3:Learner]{mlr3::Learner}}

\item{\code{resampling}}{\link[mlr3:Resampling]{mlr3::Resampling}}

\item{\code{measures}}{list of \link[mlr3:Measure]{mlr3::Measure}}

\item{\code{terminator}}{\link[mlr3tuning:Terminator]{mlr3tuning::Terminator}}

\item{\code{bmr}}{\link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n_evals}}{Number of evaluations.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FSelectInstance$new()}}
\item \href{#method-format}{\code{FSelectInstance$format()}}
\item \href{#method-print}{\code{FSelectInstance$print()}}
\item \href{#method-eval_batch}{\code{FSelectInstance$eval_batch()}}
\item \href{#method-fselect_objective}{\code{FSelectInstance$fselect_objective()}}
\item \href{#method-archive}{\code{FSelectInstance$archive()}}
\item \href{#method-best}{\code{FSelectInstance$best()}}
\item \href{#method-best_by_batch}{\code{FSelectInstance$best_by_batch()}}
\item \href{#method-clone}{\code{FSelectInstance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create new \code{FSelectInstance} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$new(task, learner, resampling, measures, terminator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{\link[mlr3:Task]{mlr3::Task}}

\item{\code{learner}}{\link[mlr3:Learner]{mlr3::Learner}}

\item{\code{resampling}}{\link[mlr3:Resampling]{mlr3::Resampling}}

\item{\code{measures}}{list of \link[mlr3:Measure]{mlr3::Measure}}

\item{\code{terminator}}{\link{Terminator}}

\item{\code{bmr}}{\link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}
Stores all evaluated \link[mlr3:ResampleResult]{mlr3::ResampleResult}s when evaluating feature combinations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{FSelectInstance} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\subsection{Method \code{format()}}{
Format method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_batch"></a>}}
\subsection{Method \code{eval_batch()}}{
Evaluates all feature combinations in \code{states} through resampling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$eval_batch(states)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{states}}{(\code{matrix}) Each row represents a 0/1 encoded feature combination.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fselect_objective"></a>}}
\subsection{Method \code{fselect_objective()}}{
Evaluates a feature combination \code{x} and returns a scalar objective value,
where the return value is negated if the measure is maximized.
This method is useful for feature selection algorithms that take a objective function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$fselect_objective(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{numeric}
0/1 encoded feature combination}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-archive"></a>}}
\subsection{Method \code{archive()}}{
Returns a table of contained resample results with corresponding feature combinations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$archive()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \link[data.table:data.table]{data.table::data.table} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-best"></a>}}
\subsection{Method \code{best()}}{
Queries the \link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult} for the \code{n} best feature combinations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$best(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{(\code{ìnteger})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link[data.table:data.table]{data.table::data.table} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-best_by_batch"></a>}}
\subsection{Method \code{best_by_batch()}}{
Queries the \link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult} for the \code{n} best feature combinations in each batch.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$best_by_batch(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{(\code{ìnteger})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link[data.table:data.table]{data.table::data.table} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
