% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSelectInstance.R
\name{FSelectInstance}
\alias{FSelectInstance}
\title{FSelectInstance Class}
\description{
Specifies a general feature selection scenario.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task}}{\link[mlr3:Task]{mlr3::Task}}

\item{\code{learner}}{\link[mlr3:Learner]{mlr3::Learner}}

\item{\code{resampling}}{\link[mlr3:Resampling]{mlr3::Resampling}}

\item{\code{measures}}{list of \link[mlr3:Measure]{mlr3::Measure}}

\item{\code{terminator}}{\link[mlr3tuning:Terminator]{mlr3tuning::Terminator}}

\item{\code{bmr}}{\link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n_evals}}{Number of evaluations.}

\item{\code{result}}{Result of the feature selection i.e. the optimal feature combination and its estimated performances.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FSelectInstance$new()}}
\item \href{#method-format}{\code{FSelectInstance$format()}}
\item \href{#method-print}{\code{FSelectInstance$print()}}
\item \href{#method-eval_batch}{\code{FSelectInstance$eval_batch()}}
\item \href{#method-fselect_objective}{\code{FSelectInstance$fselect_objective()}}
\item \href{#method-archive}{\code{FSelectInstance$archive()}}
\item \href{#method-optimization_path}{\code{FSelectInstance$optimization_path()}}
\item \href{#method-best}{\code{FSelectInstance$best()}}
\item \href{#method-assign_result}{\code{FSelectInstance$assign_result()}}
\item \href{#method-clone}{\code{FSelectInstance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create new \code{FSelectInstance} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$new(task, learner, resampling, measures, terminator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{\link[mlr3:Task]{mlr3::Task}}

\item{\code{learner}}{\link[mlr3:Learner]{mlr3::Learner}}

\item{\code{resampling}}{\link[mlr3:Resampling]{mlr3::Resampling}}

\item{\code{measures}}{list of \link[mlr3:Measure]{mlr3::Measure}}

\item{\code{terminator}}{\link{Terminator}}

\item{\code{bmr}}{\link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}
Stores all evaluated \link[mlr3:ResampleResult]{mlr3::ResampleResult}s when evaluating feature combinations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{FSelectInstance} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\subsection{Method \code{format()}}{
Format method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$format()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_batch"></a>}}
\subsection{Method \code{eval_batch()}}{
Evaluates all feature combinations in \code{states} through resampling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$eval_batch(states)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{states}}{(\code{matrix}) Each row represents a 0/1 encoded feature combination.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fselect_objective"></a>}}
\subsection{Method \code{fselect_objective()}}{
Evaluates a feature combination \code{x} and returns a scalar objective value,
where the return value is negated if the measure is maximized.
This method is useful for feature selection algorithms that take a objective function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$fselect_objective(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{numeric}
0/1 encoded feature combination}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{numeric(1)}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-archive"></a>}}
\subsection{Method \code{archive()}}{
Returns a table of contained resample results with corresponding feature subsets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$archive()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \link[data.table:data.table]{data.table::data.table} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimization_path"></a>}}
\subsection{Method \code{optimization_path()}}{
Queries the \link[ml3:BenchmarkResult]{ml3::BenchmarkResult} for the \code{n} best feature subsets (default is \code{1})
of the batches specified in \code{m} (default is the last batch).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$optimization_path(n = 1, m = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{(\code{integer})}

\item{\code{m}}{(\code{integer})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link[data.table:data.table]{data.table::data.table} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-best"></a>}}
\subsection{Method \code{best()}}{
Queries the \link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult} for the best \link[mlr3:ResampleResult]{mlr3::ResampleResult}
according to \code{measure} (default is the first measure in \verb{$measures})
of batch  \code{m} (default is the last batch).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$best(measure = NULL, m = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measure}}{\link[mlr3:Measure]{mlr3::Measure}}

\item{\code{m}}{(\code{ìnteger})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link[mlr3:ResampleResult]{mlr3::ResampleResult} objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assign_result"></a>}}
\subsection{Method \code{assign_result()}}{
The \link{FSelect} object writes the best found feature subset and estimated performance values here. For internal use.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$assign_result(feat, perf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feat}}{(\code{character}) Must be character vector of feature names existing in \code{task}}

\item{\code{perf}}{(\code{numeric}) Must be named numeric of performance measures, named with performance IDs, regarding all elements in \code{measures}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectInstance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
