% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSelectRandom.R
\name{FSelectRandom}
\alias{FSelectRandom}
\title{FSelectRandom Class}
\description{
Random feature selection.
}
\section{Super class}{
\code{\link[mlr3fselect:FSelect]{mlr3fselect::FSelect}} -> \code{FSelectRandom}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FSelectRandom$new()}}
\item \href{#method-clone}{\code{FSelectRandom$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelect" data-id="select">}\href{../../mlr3fselect/html/FSelect.html#method-select}{\code{mlr3fselect::FSelect$select()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create new \code{FSelectRandom} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectRandom$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{FSelectRandom} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectRandom$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
