% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSelectRandom.R
\name{FSelectRandom}
\alias{FSelectRandom}
\title{FSelectRandom Class}
\description{
Class for random feature selection.
}
\section{Parameters}{

\describe{
\item{\code{max_features}}{\code{integer(1)} Maximum number of features. By default, number of features in \link[mlr3:Task]{mlr3::Task}.}
\item{\code{batch_size}}{\code{integer(1)} Maximum number of feature combinations to try in a batch.}}

In order to support general termination criteria and parallelization,
feature combinations are evaluated in a batch-fashion of size \code{batch_size}.
Larger batches mean more is parallelized, smaller batches imply a more fine-grained checking
of termination criteria.
}

\section{Super class}{
\code{\link[mlr3fselect:FSelect]{mlr3fselect::FSelect}} -> \code{FSelectRandom}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FSelectRandom$new()}}
\item \href{#method-clone}{\code{FSelectRandom$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelect" data-id="format">}\href{../../mlr3fselect/html/FSelect.html#method-format}{\code{mlr3fselect::FSelect$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelect" data-id="print">}\href{../../mlr3fselect/html/FSelect.html#method-print}{\code{mlr3fselect::FSelect$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelect" data-id="select">}\href{../../mlr3fselect/html/FSelect.html#method-select}{\code{mlr3fselect::FSelect$select()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create new \code{FSelectRandom} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectRandom$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{FSelectRandom} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectRandom$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
