% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSelectSequential.R
\name{FSelectSequential}
\alias{FSelectSequential}
\title{FSelectSequential Class}
\description{
Sequential feature selection.
}
\section{Super class}{
\code{\link[mlr3fselect:FSelect]{mlr3fselect::FSelect}} -> \code{FSelectSequential}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FSelectSequential$new()}}
\item \href{#method-clone}{\code{FSelectSequential$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="FSelect" data-id="select">}\href{../../mlr3fselect/html/FSelect.html#method-select}{\code{mlr3fselect::FSelect$select()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create new \code{FSelectSequential} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectSequential$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{FSelectSequential} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FSelectSequential$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
