% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_filters.R
\name{mlr_filters}
\alias{mlr_filters}
\title{Dictionary of Filters}
\format{\link{R6Class} object}
\description{
A simple \link{Dictionary} storing objects of class \link{Filter}.
Each Filter has an associated help page, see \code{mlr_filters_[id]}.
}
\section{Usage}{


See \link{Dictionary}.
}

\examples{
mlr_filters$keys()
as.data.table(mlr_filters)
mlr_filters$get("mim")
}
\seealso{
Other Filter: \code{\link{FilterAUC}},
  \code{\link{FilterCMIM}}, \code{\link{FilterDISR}},
  \code{\link{FilterGainRatio}},
  \code{\link{FilterInformationGain}},
  \code{\link{FilterJMIM}}, \code{\link{FilterJMI}},
  \code{\link{FilterKruskalTest}},
  \code{\link{FilterLinearCorrelation}},
  \code{\link{FilterMIM}}, \code{\link{FilterMRMR}},
  \code{\link{FilterNJMIM}},
  \code{\link{FilterRankCorrelation}},
  \code{\link{FilterSymmetricalUncertainty}},
  \code{\link{FilterVariableImportance}},
  \code{\link{FilterVariance}}, \code{\link{Filter}}
}
\concept{Dictionary}
\concept{Filter}
