% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceEvaluator.R
\docType{data}
\name{PerformanceEvaluator}
\alias{PerformanceEvaluator}
\title{PerformanceEvaluator Class}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Evaluates feature sets with \link[mlr3:benchmark]{mlr3::benchmark} generated by \link{FeatureSelection}.
}
\section{Construction}{
\preformatted{pe = PerformanceEvaluator$new(task, learner, resampling)
}
\itemize{
\item \code{task} :: \link[mlr3:Task]{mlr3::Task}\cr
\item \code{learner} :: \link[mlr3:Learner]{mlr3::Learner}\cr
\item \code{resampling} :: \link[mlr3:Resampling]{mlr3::Resampling}\cr
}
}

\section{Fields}{

\itemize{
\item \code{task} :: \link[mlr3:Task]{mlr3::Task}\cr Stores the task.
\item \code{learner} :: \link[mlr3:Learner]{mlr3::Learner}\cr Stores the learner.
\item \code{resampling} :: \link[mlr3:Resampling]{mlr3::Resampling}\cr Stores the resampling.
\item \code{bmr} :: list of \link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}\cr Stores the benchmark results.
}
}

\section{Methods}{

\itemize{
\item \code{eval_states(states)}\cr
list of \code{integer()} -> \code{self}\cr
Evaluates feature sets.
}
}

\examples{
task = mlr3::mlr_tasks$get("iris")
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("holdout")
pe = PerformanceEvaluator$new(task, learner, resampling)
}
\concept{PerformanceEvaluator}
\keyword{datasets}
