% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Terminator.R
\name{Terminator}
\alias{Terminator}
\title{Abstract Terminator Class}
\description{
Abstract \code{Terminator} class that implements the main functionality each terminator must have. A terminator is an object that says when to stop the feature selection.
}
\section{Usage}{
\preformatted{# Construction
tm = Terminator$new()

# Public members
tm$terminated
tm$state

# Public methods
tm$update_state(pe)
tm$update_end(pe)
}
}

\section{Arguments}{

\itemize{
\item \code{pe} (\code{PerformanceEvaluator}):
\code{PerformanceEvaluator} object used to determine when to stop the tuning.
}
}

\section{Details}{

\itemize{
\item \code{$new()} creates a new object of class \link{Terminator}.
\item \code{$terminated} (\code{logical(1)}) is the termination criterion met? Updated by each call of \code{update_start()}/\code{update_end()}.
\item \code{$settings} (\code{list()}) settings that are set by the child classes to define stopping criteria.
\item \code{$state} (\code{list()}) arbitrary state of the Terminator. Gets updated with each call of \code{update_start()} and \code{update_end()}.
\item \code{$update_start()} is called in each tuning iteration before the evaluation.
\item \code{$update_end()} is called in each tuning iteration after the evaluation.
}
}

\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorPerformanceStep}}
}
\concept{Terminator}
\keyword{internal}
