% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{TerminatorClockTime}
\alias{TerminatorClockTime}
\title{Re-export of \code{TerminatorClockTime}

See \link[bbotk:TerminatorClockTime]{bbotk::TerminatorClockTime}.}
\description{
Re-export of \code{TerminatorClockTime}

See \link[bbotk:TerminatorClockTime]{bbotk::TerminatorClockTime}.
}
