% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorEvaluations.R
\name{TerminatorEvaluations}
\alias{TerminatorEvaluations}
\title{TerminatorEvaluations}
\description{
Terminator child class to terminate the feature selection if the model performance does not improve to a specified threshold in the next step.
}
\section{Usage}{
\preformatted{tm = TerminatorEvaluations$new()
}

See \link{Terminator} for a description of the interface.
}

\section{Arguments}{

\itemize{
\item \code{max_evaluations} (\code{integer(1)}):
Maximum number of function evaluations.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{TerminatorEvaluations}.

The interface is described in \link{Terminator}.
}

\examples{
t = TerminatorEvaluations$new(max_evaluations = 100)
}
\seealso{
Other Terminator: \code{\link{TerminatorPerformanceStep}},
  \code{\link{Terminator}}
}
\concept{Terminator}
