% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorNone.R
\name{TerminatorNone}
\alias{TerminatorNone}
\alias{mlr_terminators_none}
\title{Terminator that never stops.}
\description{
Mainly useful for optimization algorithms, where the stopping is inherently controlled by the algorithm itself (e.g. grid search).
}
\section{Super class}{
\code{\link[mlr3fselect:Terminator]{mlr3fselect::Terminator}} -> \code{TerminatorNone}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TerminatorNone$new()}}
\item \href{#method-is_terminated}{\code{TerminatorNone$is_terminated()}}
\item \href{#method-clone}{\code{TerminatorNone$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="Terminator" data-id="format">}\href{../../mlr3fselect/html/Terminator.html#method-format}{\code{mlr3fselect::Terminator$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3fselect" data-topic="Terminator" data-id="print">}\href{../../mlr3fselect/html/Terminator.html#method-print}{\code{mlr3fselect::Terminator$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create new \code{TerminatorNone} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorNone$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{TerminatorNone} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_terminated"></a>}}
\subsection{Method \code{is_terminated()}}{
Is always \code{FALSE}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorNone$is_terminated(instance)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{object of class \link{FSelect}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{logical(1)}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorNone$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
