% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorPerformanceStep.R
\docType{data}
\name{TerminatorPerformanceStep}
\alias{TerminatorPerformanceStep}
\title{TerminatorPerformanceStep}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Terminator}.}
\usage{
TerminatorPerformanceStep
}
\description{
Feature selection terminates depending on the performance improvement between the current and previous evaluation set.
}
\section{Construction}{
\preformatted{tm = TerminatorPerformanceStep$new(threshold)
}
\itemize{
\item \code{threshold} (`numeric(1)``)\cr
}
}

\section{Fields}{

See \link{Terminator}.
}

\section{Methods}{

See \link{Terminator}.
}

\examples{
tm = TerminatorPerformanceStep$new(threshold = 0.01)
}
\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorRuntime}}, \code{\link{Terminator}}
}
\concept{Terminator}
\keyword{datasets}
