% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorPerformanceStep.R
\name{TerminatorPerformanceStep}
\alias{TerminatorPerformanceStep}
\title{TerminatorPerformanceStep}
\description{
Terminator child class to terminate the sequential feature selection if the model performance does not improve to a specified threshold in the next step.
}
\section{Usage}{
\preformatted{tm = TerminatorPerformanceStep$new(threshold)
}

See \link{Terminator} for a description of the interface.
}

\section{Arguments}{

\itemize{
\item \code{pe} (\code{[PerformanceEvaluator]})
\item \code{threshold} (`numeric(1)``):
The feature selection is terminated if the performance improvement between two steps is less than the threshold.
\item \code{max_features} (\code{integer(1)})
Maximum number of features. Only effects \link{FeatureSelectionForward}.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{TerminatorPerformanceStep}.

The interface is described in \link{Terminator}.
}

\examples{
task = mlr3::mlr_tasks$get("iris")
learner = mlr3::mlr_learners$get("classif.rpart")
resampling = mlr3::mlr_resamplings$get("holdout")
pe = PerformanceEvaluator$new(task, learner, resampling)
tm = TerminatorPerformanceStep$new(pe, threshold = 0.01, max_features = 2)
}
\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{Terminator}}
}
\concept{Terminator}
