% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{TerminatorPerfReached}
\alias{TerminatorPerfReached}
\title{Re-export of \code{TerminatorPerfReached}

See \link[bbotk:TerminatorPerfReached]{bbotk::TerminatorPerfReached}.}
\description{
Re-export of \code{TerminatorPerfReached}

See \link[bbotk:TerminatorPerfReached]{bbotk::TerminatorPerfReached}.
}
