% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorRuntime.R
\docType{data}
\name{TerminatorRuntime}
\alias{TerminatorRuntime}
\title{TerminatorRuntime Class}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Terminator}.}
\usage{
TerminatorRuntime
}
\description{
Feature selection terminates depeding on the runtime.
}
\section{Construction}{
\preformatted{tm = TerminatorRuntime$new(max_time, time_unit)
}
\itemize{
\item \code{max_time} :: \code{numeric(1)}\cr Maximum allowed runtime.
\item \code{units} :: \code{character(1)}\cr Unit of the maximum allowed runtime in \code{secs}, \code{mins}, \code{hours}, \code{days}, or \code{weeks}.
}
}

\section{Fields}{

See \link{Terminator}.
}

\section{Methods}{

See \link{Terminator}.
}

\examples{
tm = TerminatorRuntime$new(max_time = 5, units = "secs")
}
\seealso{
Other Terminator: \code{\link{TerminatorEvaluations}},
  \code{\link{TerminatorPerformanceStep}},
  \code{\link{Terminator}}
}
\concept{Terminator}
\keyword{datasets}
